+++ src/backend/community-service/src/main/java/harmony/communityservice/guild/channel/dto/DeleteChannelRequest.java
@@ -3,7 +3,7 @@ package harmony.communityservice.guild.channel.dto;
 import harmony.communityservice.common.dto.CommonRequest;
 import jakarta.validation.constraints.NotNull;
 
-public record DeleteChannelRequest(@NotNull Long channelId, @NotNull Long guildId, @NotNull Long userId)
+public record DeleteChannelRequest(@NotNull Integer channelId, @NotNull Long guildId, @NotNull Long userId)
         implements CommonRequest {
 
     @Override
