+++ src/backend/community-service/src/main/java/harmony/communityservice/guild/channel/dto/SearchChannelResponse.java
@@ -4,6 +4,6 @@ import harmony.communityservice.guild.domain.ChannelType;
 import lombok.Builder;
 
 @Builder(toBuilder = true)
-public record SearchChannelResponse(Long channelId, Long guildId, Long categoryId, String channelName,
+public record SearchChannelResponse(Integer channelId, Long guildId, Integer categoryId, String channelName,
                                     ChannelType channelType) {
 }
