+++ src/backend/community-service/src/main/java/harmony/communityservice/guild/channel/service/query/ChannelQueryService.java
@@ -1,10 +1,10 @@
 package harmony.communityservice.guild.channel.service.query;
 
-import harmony.communityservice.guild.channel.dto.SearchChannelResponse;
 import harmony.communityservice.common.dto.SearchParameterMapperRequest;
+import harmony.communityservice.guild.channel.dto.SearchChannelResponse;
 import java.util.Map;
 
 public interface ChannelQueryService {
 
-    Map<Long, SearchChannelResponse> searchMapByGuildId(SearchParameterMapperRequest searchParameterMapperRequest);
+    Map<Integer, SearchChannelResponse> searchMapByGuildId(SearchParameterMapperRequest searchParameterMapperRequest);
 }
